Separate Entry from (Entry) Draft
In the initial SAO API design we wanted to merge what is
conceptually defined as an Entry ("submittion" in CMS) and a
Draft ("user-test" in CMS), since both share an almost identical
entity definition into the database, with the only exception that
the latter one is not being referenced by a Result tuple.

Unfortunately, merging Entry and Draft together carried several
compromises in the API implementation. So, I'm basically stepping
back the initial intention.