Add Result database abstraction
Since the Score resource will be embeded to the Result one, the
storage queries had to reflect it correctly.

Worth mentioning that the most difficult part was querying ONLY
the results with the MAX value without using aggregation functions
or sub-queries. Kudos to "axiac" in Stack Overflow I owe you one.