Add DraftResult database abstraction
The Draft Result entity definition was quite similar to the Result
one with the main difference on how an "Execution" was referenced,
since later have many of them while the former only one.

As a side note, I still wonder the reasoning behind creating the
Draft Result output as a PostgreSQL FSObject (AKA blob) instead
of a nchar value.